requirements.txt
nextcord==2.0.0b3
pysqlite3==0.4.7
aiosqlite==0.17.0
python-dotenv==0.20.0
cassiopeia==5.0.1
motor==3.0.0